image/requirements.txt
autograd==1.6.2
Cython==3.0.10
deepinv==0.2.0
diffeqtorch==1.0.0
gmpy2
h5py
jax
lightning
matplotlib
ninja
numdifftools
numpy
opencv-python
pandas
Pillow
POT
pykeops
pytorch-ignite
pytorch-lightning
scikit-image
scikit-learn
scipy
sympy
tensorboard
tensorboard-data-server
toolz
torch
torchdiffeq
torchmetrics
torchvision
tqdm
urllib3
ml_collections
datasets
torchmetrics[image]
tyro